Rename Parser and Session
SDP.Session is SDP now

SDP is SDP.Parser now

Add parse and serialize functions to SDP, delegate them to SDP.Serializer and SDP.Parser
